+++ lib/WorkflowMain.groovy
@@ -21,40 +21,11 @@ class WorkflowMain {
             "  https://github.com/${workflow.manifest.name}/blob/master/CITATIONS.md"
     }
 
-    //
-    // Generate help string
-    //
-    public static String help(workflow, params) {
-        def command = "nextflow run ${workflow.manifest.name} --input samplesheet.csv -profile docker"
-        def help_string = ''
-        help_string += NfcoreTemplate.logo(workflow, params.monochrome_logs)
-        help_string += NfcoreSchema.paramsHelp(workflow, params, command)
-        help_string += '\n' + citation(workflow) + '\n'
-        help_string += NfcoreTemplate.dashedLine(params.monochrome_logs)
-        return help_string
-    }
-
-    //
-    // Generate parameter summary log string
-    //
-    public static String paramsSummaryLog(workflow, params) {
-        def summary_log = ''
-        summary_log += NfcoreTemplate.logo(workflow, params.monochrome_logs)
-        summary_log += NfcoreSchema.paramsSummaryLog(workflow, params)
-        summary_log += '\n' + citation(workflow) + '\n'
-        summary_log += NfcoreTemplate.dashedLine(params.monochrome_logs)
-        return summary_log
-    }
 
     //
     // Validate parameters and print summary to screen
     //
     public static void initialise(workflow, params, log) {
-        // Print help to screen if required
-        if (params.help) {
-            log.info help(workflow, params)
-            System.exit(0)
-        }
 
         // Check that keys for reference databases are valid
         if (params.dada_ref_taxonomy && !params.skip_taxonomy && !params.skip_dada_taxonomy) {
@@ -66,20 +37,15 @@ class WorkflowMain {
         if (params.qiime_ref_taxonomy && !params.skip_taxonomy && !params.classifier) {
             qiimereftaxonomyExistsError(params, log)
         }
+        if (params.sintax_ref_taxonomy && !params.skip_taxonomy) {
+            sintaxreftaxonomyExistsError(params, log)
+        }
 
         // Print workflow version and exit on --version
         if (params.version) {
             String workflow_version = NfcoreTemplate.version(workflow)
             log.info "${workflow.manifest.name} ${workflow_version}"
             System.exit(0)
-        }
-
-        // Print parameter summary log to screen
-        log.info paramsSummaryLog(workflow, params)
-
-        // Validate workflow parameters via the JSON schema
-        if (params.validate_params) {
-            NfcoreSchema.validateParameters(workflow, params, log)
         }
 
         // Check that a -profile or Nextflow config has been provided to run the pipeline
@@ -129,6 +95,19 @@ class WorkflowMain {
                 "  QIIME2 reference database '${params.qiime_ref_taxonomy}' not found in any config file provided to the pipeline.\n" +
                 "  Currently, the available reference taxonomy keys for `--qiime_ref_taxonomy` are:\n" +
                 "  ${params.qiime_ref_databases.keySet().join(", ")}\n" +
+                "==================================================================================="
+            Nextflow.error(error_string)
+        }
+    }
+    //
+    // Exit pipeline if incorrect --qiime_ref_taxonomy key provided
+    //
+    private static void sintaxreftaxonomyExistsError(params, log) {
+        if (params.sintax_ref_databases && params.sintax_ref_taxonomy && !params.sintax_ref_databases.containsKey(params.sintax_ref_taxonomy)) {
+            def error_string = "=============================================================================\n" +
+                "  SINTAX reference database '${params.sintax_ref_taxonomy}' not found in any config files provided to the pipeline.\n" +
+                "  Currently, the available reference taxonomy keys for `--sintax_ref_taxonomy` are:\n" +
+                "  ${params.sintax_ref_databases.keySet().join(", ")}\n" +
                 "==================================================================================="
             Nextflow.error(error_string)
         }